+++ doorstop/cli/test/test_all.py
@@ -298,6 +298,11 @@ class TestLogging(unittest.TestCase):  # pylint: disable=R0904
         self.assertIs(None, main(['-vvvv']))
 
     def test_verbose_5(self):
-        """Verify verbose level 5 cannot be set."""
+        """Verify verbose level 5 can be set."""
         self.assertIs(None, main(['-vvvvv']))
-        self.assertEqual(4, common.VERBOSITY)
+        self.assertEqual(5, common.VERBOSITY)
+
+    def test_verbose_6(self):
+        """Verify verbose level 6 cannot be set."""
+        self.assertIs(None, main(['-vvvvvv']))
+        self.assertEqual(5, common.VERBOSITY)