+++ doorstop/core/test/test_document.py
@@ -87,12 +87,12 @@ class TestDocument(unittest.TestCase):  # pylint: disable=R0904
 
     def test_str(self):
         """Verify documents can be converted to strings."""
-        common.VERBOSITY = 1
+        common.VERBOSITY = 2
         self.assertEqual("REQ", str(self.document))
 
     def test_str_verbose(self):
         """Verify documents can be converted to strings in verbose mode."""
-        common.VERBOSITY = 2
+        common.VERBOSITY = 3
         relpath = os.path.relpath(self.document.path, self.document.root)
         text = "REQ (@{}{})".format(os.sep, relpath)
         self.assertEqual(text, str(self.document))