+++ doorstop/settings.py
@@ -6,8 +6,8 @@ import logging
 
 # Logging settings
 DEFAULT_LOGGING_FORMAT = "%(message)s"
-VERBOSE_LOGGING_FORMAT = "%(levelname)s: %(message)s"
-VERBOSE2_LOGGING_FORMAT = "%(levelname)s: %(module)s:%(lineno)d: %(message)s"
+VERBOSE_LOGGING_FORMAT = "[%(levelname)-8s] %(message)s"
+VERBOSE2_LOGGING_FORMAT = "[%(levelname)-8s] (%(module)-8s @%(lineno)4d) %(message)s"
 DEFAULT_LOGGING_LEVEL = logging.WARNING
 VERBOSE_LOGGING_LEVEL = logging.INFO
 VERBOSE2_LOGGING_LEVEL = logging.DEBUG